Move extension to separate file
