+++ Wire-iOS/Sources/UserInterface/Settings/CellDescriptors/SettingsExternalScreenCellDescriptor.swift
@@ -154,25 +154,3 @@ class SettingsExternalScreenCellDescriptor: SettingsExternalScreenCellDescriptor
         return self.presentationAction()
     }
 }
-
-extension UITableViewCell {
-    func showDisclosureIndicator() {
-        if #available(iOS 13.0, *) {
-            let image = UIImage(systemName: "chevron.right")
-            let accessory = UIImageView(frame: CGRect(x: 0, y: 0, width: (image?.size.width)!, height: (image?.size.height)!))
-            accessory.image = image
-
-            accessory.tintColor = UIColor.white
-            accessoryView = accessory
-        } else {
-            accessoryType = .disclosureIndicator
-        }
-    }
-    func hideDisclosureIndicator() {
-        if #available(iOS 13.0, *) {
-            accessoryView = nil
-        } else {
-            accessoryType = .none
-        }
-    }
-}
